LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/allow-scripts-flag-changing-2-expected.txt

CONSOLE MESSAGE: line 45: TypeError: Argument 1 ('node') to Node.appendChild must be an instance of Node

Harness Error (FAIL), message = TypeError: Argument 1 ('node') to Node.appendChild must be an instance of Node
